refactor(lnurlpay): align usage via prompt and popup #1216
- split component into functionality and form
